+++ src/app.js
@@ -522,10 +522,28 @@ export class App {
         Define the scissor box.
 
         @method
+        @param {Number} x Horizontal position of the scissor box.
+        @param {Number} y Vertical position of the scissor box.
+        @param {Number} width Width of the scissor box.
+        @param {Number} height Height of the scissor box.
         @return {App} The App object.
     */
     scissor(x, y, width, height) {
         this.gl.scissor(x, y, width, height);
+
+        return this;
+    }
+
+    /**
+        Set the scale and units used.
+
+        @method
+        @param {Number} factor Scale factor used to create a variable depth offset for each polygon.
+        @param {Number} units Constant depth offset.
+        @return {App} The App object.
+    */
+    polygonOffset(factor, units) {
+        this.gl.polygonOffset(factor, units);
 
         return this;
     }
@@ -835,12 +853,17 @@ export class App {
 
         @method
         @param {GLEnum} type The data type stored in the index buffer.
-        @param {number} itemSize Number of elements per primitive.
+        @param {number} [itemSize=3] Number of elements per primitive.
         @param {ArrayBufferView} data Index buffer data.
         @param {GLEnum} [usage=STATIC_DRAW] Buffer usage.
         @return {VertexBuffer} New VertexBuffer object.
     */
     createIndexBuffer(type, itemSize, data, usage) {
+        if (ArrayBuffer.isView(itemSize)) {
+            usage = data;
+            data = itemSize;
+            itemSize = 3;
+        }
         return new VertexBuffer(this.gl, this.state, type, itemSize, data, usage, true);
     }