+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## v0.17.4
+- App.polygonOffset method
+
+## v0.17.3
+
+- VertexArray.vertexAttributeBuffer, VertexArray.instanceAttributeBuffer `size`, `offset` and `stride` options will affect item count.
+- App.createIndexBuffer no longer needs an item count (defaults to 3).
+
+## v0.17.2
+
+- VertexBuffer.data can take an offset.
+
 ## v0.17.1
 
 - Support for changing framebuffer binding points.